some shuffling, add a `-q` flag (to suppress file/line output)
